// client/src/main/cv-viewer/src/app/api/model/cvPermissions.ts
/**
 * CV Viewer
 * CV Viewer.
 *
 * OpenAPI spec version: 1.0.0
 * Contact: neilje123@gmail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface CvPermissions { 
    cvName?: string;
    cvOwner?: string;
    cvUsers?: Array<string>;
}